Omit sha256 hash unless requested
Block0-Block1 images should only contain the sha256 hash
if they are from older builds that explicitly request the
hash as part of block0. Newer builds that have moved on to
sha384 should not contain the sha256 hash, but 0xff instead.

Tested: build an image with the block1/b0_sig/hashalg set to
        sha384 and see that the sha256 hash is set to 0xff.

Change-Id: Ic2368808e17e7a6318f5126bf416266a6cd40a44
Signed-off-by: Vernon Mauery <vernon.mauery@intel.com>