Add OEM data for CPLD images
CPLD images can now have an optional OEM data field of 16 bytes
following the SVN field. The oemdata can to be either a 16-byte
ASCII string or a hex value. If the data starts with 0x, it will be
interpreted as a hex value.

Tested: manually run intel-pfr-signing-utility with a CPLD config
    that contains the oemdata field. Also, tested with bad values
    to see that they were rejected by the input checking.

Signed-off-by: Vernon Mauery <vernon.mauery@intel.com>